uconn-sentiment-backend/requirements.txt
pandas==2.0.0
transformers==4.39.3
torch==2.2.0
nltk==3.8.1
scikit-learn==1.4.2
openpyxl==3.1.2
google-cloud-datastore==2.19.0
beautifulsoup4==4.12.3google-cloud-datastore==2.19.0
pyyaml==6.0.1